+++ canvas/getWidth.go
@@ -5,6 +5,6 @@ package canvas
 //     The width property returns the width of an ImageData object, in pixels.
 //     Tip: Look at createImageData(), getImageData(), and putImageData() to learn more about the ImageData object.
 //     JavaScript syntax: imgData.width;
-func (el *Canvas) GetWidth() int {
-	return el.SelfContext.Get("width").Int()
+func (el *Canvas) GetWidth() float64 {
+	return el.SelfContext.Get("width").Float()
 }
