+++ canvas/setX.go
@@ -1,5 +1,5 @@
 package canvas
 
-func (el *Canvas) SetX(x int) {
+func (el *Canvas) SetX(x float64) {
 	el.SelfContext.Set("x", x)
 }
